RpMemFile: Fixed issues reading out of bounds.
RpImageLoader reads the first 256 bytes of the image to determine what
kind of file it is. odd-width.16color.CI4.png and happy-mac.mono.odd-size.png
are both less than 256 bytes. However, the boundary check for RpMemFile
was using unsigned arithmetic, which prevented it from detecting an
out-of-bounds read in this case.

Cast the size_t values to int64_t to fix this issue.

This caused an SEH exception when running RpImageFormatTest on Windows XP
when using libpng instead of GDI+:

[ RUN      ] happy_mac_mono_png/RpPngFormatTest.loadTest/happy_mac_mono_odd_size_png
unknown file: error: SEH exception with code 0xc0,000,005 thrown in the test body.
[  FAILED  ] happy_mac_mono_png/RpPngFormatTest.loadTest/happy_mac_mono_odd_size_png, where GetParam() = happy-mac.mono.odd-size.png (10 ms)

+++ src/libromdata/file/RpMemFile.cpp
@@ -149,7 +149,8 @@ size_t RpMemFile::read(void *ptr, size_t size)
 	const uint8_t *buf = reinterpret_cast<const uint8_t*>(m_buf);
 
 	// Check if size is in bounds.
-	if (m_pos > m_size - size) {
+	// NOTE: Need to use a signed comparison here.
+	if ((int64_t)m_pos > ((int64_t)m_size - (int64_t)size)) {
 		// Not enough data.
 		// Copy whatever's left in the buffer.
 		size = m_size - m_pos;
